+++ NGitLab.Mock/Clients/MergeRequestCommitClient.cs
@@ -6,10 +6,10 @@ namespace NGitLab.Mock.Clients;
 
 internal sealed class MergeRequestCommitClient : ClientBase, IMergeRequestCommitClient
 {
-    private readonly int _projectId;
-    private readonly int _mergeRequestIid;
+    private readonly long _projectId;
+    private readonly long _mergeRequestIid;
 
-    public MergeRequestCommitClient(ClientContext context, int projectId, int mergeRequestIid)
+    public MergeRequestCommitClient(ClientContext context, long projectId, long mergeRequestIid)
         : base(context)
     {
         _projectId = projectId;
